requirements.txt
requests
openpyxl
pandas
python-dotenv
beautifulsoup4
lxml